backend/requirements.txt
flask
flask_sqlalchemy
flask_cors 
psycopg2 
python-dotenv 